% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chibi.anova.R
\name{chibi.anova}
\alias{chibi.anova}
\title{Plots results from aov output}
\usage{
chibi.anova(
  myanova = NULL,
  pval_thres = 0.05,
  legend_proportion_size = 2,
  y_vjust = 1,
  size_axis_text = 20,
  size_axis_title = 30,
  size_legend_text = 20,
  size_title_text = 30,
  size_ticks_x = 2.5,
  size_ticks_y = 2.5,
  font_family = "Arial",
  aspect.ratio = 3,
  size_panel_border = 0.1
)
}
\description{
Uses the aov table contained in out by aov
}
\examples{
chibi.anova()
}
\keyword{chibi,anova,plot}
