% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chibi.boxplot.R
\name{chibi.boxplot}
\alias{chibi.boxplot}
\title{Creates a boxplot ussing ggplot2 using variables in a metadata (Map) structure.}
\usage{
chibi.boxplot(Map = Map, x_val = NULL, y_val = NULL, col_val = NULL,
  shape_val = NULL, mypch = 21, size = 10, alpha = 0.7, stroke = 0.5)
}
\description{
The boxplot can represent up to three variables from the Map object
}
\examples{
chibi.boxplot()
}
\keyword{boxplot}
