% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chibi.cap.R
\name{chibi.cap}
\alias{chibi.cap}
\title{Plots results from oh.cap}
\usage{
chibi.cap(list_ohpco = NULL, col_val = NULL, comp_a = NULL,
  comp_b = NULL, mypch = 22, size = 25, alpha = 0.7, stroke = 1.5)
}
\description{
Uses the structure return from oh.cap to automatically plot a scatterplot colored by variables from the Map structured used to compute oh.cap
}
\examples{
chibi.cap()
}
\keyword{cap}
