% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chibi.comparecluster.R
\name{chibi.comparecluster}
\alias{chibi.comparecluster}
\title{Perform GO enrichment comparing clusters}
\usage{
chibi.comparecluster(
  melted = NULL,
  paleta = "viridis::plasma",
  showCategory = 20,
  cutoff_redundancy = 0.7
)
}
\description{
Automatically perform the comparison at 3 levels
}
\examples{
chibi.comparecluster()
}
\keyword{chibi,GO,ggplot}
