% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chibi.heatmap.R
\name{chibi.heatmap}
\alias{chibi.heatmap}
\title{Plot a  heatmap}
\usage{
chibi.heatmap(
  Tab = NULL,
  dist_method_rows = "pearson",
  dist_method_cols = "euclidean",
  hclust_method_rows = "complete",
  hclust_method_cols = "ward.D2",
  k_rows = 8,
  k_cols = 4,
  palette_heatmap = "pals::kovesi.diverging_bwr_55_98_c37",
  range_fill_heatmap = 1,
  size_strip_text_row = 10,
  size_strip_text_col = 0,
  size_axis_text_col = 10,
  size_axis_text_row = 0,
  axis_ticks_row = F,
  size_legend_text = 10,
  size_legend_title = 12,
  size_dendrogram = 0.3,
  panel_border_heatmap = 0.3,
  panel_spacing = 0,
  font_family = "Helvetica",
  font_face = "plain",
  legend_position = "right",
  egg_heights = c(0.5, 1),
  egg_widths = c(0.2, 1)
)
}
\description{
Automatically clusters, and plot
}
\examples{
chibi.heatmap()
}
\keyword{chibi,heatmap,ggplot}
