% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chibi.pca.R
\name{chibi.pca}
\alias{chibi.pca}
\title{Plots results from oh.pca
Uses the variables in the metadata (Map) to color or shape the ggplot based graph.}
\usage{
chibi.pca(list_ohpca = NULL, col_val = NULL, shape_val = NULL,
  comp_a = "PC1", comp_b = "PC2", mypch = 21, size = 25,
  alpha = 1, stroke = 1.5, col_shape_background = "white",
  alpha_shape_background = 0, lines_zero = TRUE,
  size_axis_line = 0.5, type_axis_line = "longdash",
  ratio_size_shape_background = 1.3, y_vjust = 0.5, x_hjust = 0.5,
  size_axis_text = 20, size_axis_title = 30, size_legend_text = 20,
  size_title_text = 30, legend_proportion_size = 2,
  size_lines_panel = 0, size_panel_border = 1, font_family = "Arial")
}
\description{
Plots results from oh.pca
Uses the variables in the metadata (Map) to color or shape the ggplot based graph.
}
\examples{
chibi.pca()
}
\keyword{pca}
