% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chibi.pco.R
\name{chibi.pco}
\alias{chibi.pco}
\title{Plots results from oh.pco}
\usage{
chibi.pco(list_ohpco = NULL, col_val = NULL, shape_val = NULL,
  comp_a = "PCo1", comp_b = "PCo2", mypch = 22, size = 25,
  alpha = 0.7, stroke = 1.5)
}
\description{
Uses the variables in the metadata (Map) to color or shape the ggplot based graph.
}
\examples{
chibi.pco()
}
\keyword{pco}
