% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAD.R
\name{MAD}
\alias{MAD}
\title{Performs MAD rooting of a phylogenetic tree}
\usage{
MAD(unrooted_newick, output_mode)
}
\description{
Performs MAD rooting of a phylogenetic tree
}
\examples{
MAD()
}
\keyword{MAD,}
\keyword{phylogenetics}
\keyword{tree,}
