% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh.cap.R
\name{oh.cap}
\alias{oh.cap}
\title{Performs Constrained Analysis of Principal Coordinates based on vegan capscale}
\usage{
oh.cap(Tab = NULL, Map = NULL, formula = formula, distfun = distfun,
  perms = 5000, sqrt = TRUE)
}
\description{
Performs CAP using capscale and automatically computes significance of the model and terms based on permutation. As well automatically extract constrained and unconstrained variances.
}
\examples{
oh.cap()
}
\keyword{cap}
