% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh.ggsave.R
\name{oh.ggsave.svg}
\alias{oh.ggsave.svg}
\title{Saves a ggplot object in a publication ready high resolution formar}
\usage{
oh.ggsave.svg(ggobject = p, outname = "", outdir = "figures/",
  width = 20, height = 15, device = "svg", dpi = 600)
}
\description{
Saves a ggplot object in a publication ready high resolution formar
}
\examples{
oh.ggsave()
}
\keyword{ggsave}
