% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh.hotspot.byindex.R
\name{oh.hotspot.byindex}
\alias{oh.hotspot.byindex}
\title{Given a list of numbers (genomic indices) finds indices that are physically closed in a genome.
Requires a threshold value of window to define distance between indices. Default 10}
\usage{
oh.hotspot.byindex(thres = 10, indices = NULL, num_hotspot = 0)
}
\description{
Given a list of numbers (genomic indices) finds indices that are physically closed in a genome.
Requires a threshold value of window to define distance between indices. Default 10
}
\examples{
oh.hotspot.byindex()
}
\keyword{hotspot,indices,index}
