% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh.hotspot.finder.R
\name{oh.hotspot.finder}
\alias{oh.hotspot.finder}
\title{Takes a gff df read by oh.read.gff and a list of gene_oids and identify hotspots where the
passed gene_oids are in physical proximity. Returns a dataframe with the identified hotspots.}
\usage{
oh.hotspot.finder(df_gff = NULL, gene_oids = NULL, thres = 10)
}
\description{
Takes a gff df read by oh.read.gff and a list of gene_oids and identify hotspots where the
passed gene_oids are in physical proximity. Returns a dataframe with the identified hotspots.
}
\examples{
oh.hotspot.finderx()
}
\keyword{hotspot,indices,wrapper,gff}
