% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh.MAD.R
\name{oh.MAD}
\alias{oh.MAD}
\title{Wrapper for MAD function}
\usage{
oh.MAD(tree = NULL, outfile = "rooted_tree.MAD.newick")
}
\description{
Wrapper for MAD function
}
\examples{
oh.MAD()
}
\keyword{MAD,}
\keyword{oh}
\keyword{phylogenetics,}
\keyword{tree,}
\keyword{wrapper,}
