% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh.pca.R
\name{oh.pca}
\alias{oh.pca}
\title{Performs Principal Component Analysis based on prcomp command}
\usage{
oh.pca(Tab = Tab, Map = Map, distfun = distfun, retx = TRUE,
  center = FALSE, scale = FALSE, id_var = "Sample_Id")
}
\description{
Two fundamental structures needed. The abundance matrix (Tab) and metadata (Map)
}
\examples{
oh.pca()
}
\keyword{pca}
