% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh.pco.R
\name{oh.pco}
\alias{oh.pco}
\title{Performs Principal Coordinate Analysis based on cmdscale command}
\usage{
oh.pco(Tab = Tab, Map = Map, ndim = 3, eig = T,
  distfun = distfun, id_var = "Id")
}
\description{
Two fundamental structures needed. The abundance matrix (Tab) and metadata (Map)
}
\examples{
oh.pco()
}
\keyword{pco}
