% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh.read.gff.R
\name{oh.read.gff}
\alias{oh.read.gff}
\title{Read a gff3 file (IMG Data)  and converts into df 
Aditionally extracts the gene_id and locus_tag and place them into a new column}
\usage{
oh.read.gff(file = NULL)
}
\description{
Read a gff3 file (IMG Data)  and converts into df 
Aditionally extracts the gene_id and locus_tag and place them into a new column
}
\examples{
oh.read.gff()
}
\keyword{gff,IMG,JGI,read}
