% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh.wrapper.hotspot.profiler.R
\name{oh.wrapper.hotspot.profiler}
\alias{oh.wrapper.hotspot.profiler}
\title{Wrapper that takes a gff dataframe, a dataframe with gene_ids and hmm ids (oh.profiler.pl output)and returns
two structures. One is a modified gff with the hotspots localized and another, a df summarizing the
characteristics of the hotspots identified.}
\usage{
oh.wrapper.hotspot.profiler(df_gff = NULL, df_profiler = NULL, threshold = 10)
}
\description{
Wrapper that takes a gff dataframe, a dataframe with gene_ids and hmm ids (oh.profiler.pl output)and returns
two structures. One is a modified gff with the hotspots localized and another, a df summarizing the
characteristics of the hotspots identified.
}
\examples{
oh.wrapper.hotspot.profiler()
}
\keyword{hotspot,indices,wrapper,gff}
