% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chibi.cap.R
\name{chibi.cap}
\alias{chibi.cap}
\title{Plots results from oh.cap}
\usage{
chibi.cap(
  list_ohpco = NULL,
  col_val = NULL,
  shape_val = NULL,
  comp_a = NULL,
  comp_b = NULL,
  mypch = 21,
  size = 25,
  alpha = 1,
  stroke = 1.5,
  col_shape_background = "white",
  alpha_shape_background = 0,
  lines_zero = TRUE,
  size_axis_line = 0.5,
  type_axis_line = "longdash",
  ratio_size_shape_background = 1.3,
  y_vjust = 0.5,
  x_hjust = 0.5,
  size_axis_text = 20,
  size_axis_title = 30,
  size_legend_text = 20,
  size_title_text = 30,
  legend_proportion_size = 2,
  size_lines_panel = 0,
  size_panel_border = 1,
  font_family = "Arial"
)
}
\description{
Uses the structure return from oh.cap to automatically plot a scatterplot colored by variables from the Map structured used to compute oh.cap
}
\examples{
chibi.cap()
}
\keyword{cap,chibi,plot}
