% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chibi.ggcor.R
\name{chibi.ggcor}
\alias{chibi.ggcor}
\title{Plot a correlation matrix as a heatmap}
\usage{
chibi.ggcor(
  Tab = NULL,
  hclust_method = "ward.D2",
  cor.method = "pearson",
  display.values = TRUE,
  display.significance = TRUE,
  p.adjust.method = "holm",
  p.adj.thres = 0.1,
  size_r_display = 4
)
}
\description{
Automatically clusters, calculate p.values, adjust p.values and plots
}
\examples{
chibi.ggcor()
}
\keyword{chibi,correlation,ggplot}
