% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chibi.phylogram.R
\name{chibi.phylogram}
\alias{chibi.phylogram}
\title{Plots a proportion barplot(phylogram)}
\usage{
chibi.phylogram(Tab, Map = NULL, facet = NULL, colname = "Sample",
  variable.name = "Taxon", value.name = "Abundance", scales = "free_x",
  space = "free_x", nrow.legend = 20, ntaxa = NULL,
  other_name = "other", mformula = NULL, funsum = "mean", y_vjust = 0.5,
  size_axis_text = 20, size_axis_title = 30, size_legend_text = 20,
  size_strip_text = 10)
}
\description{
Plots a proportion barplot(phylogram)
}
\examples{
chibi.phylogram()
}
\keyword{phylogram}
