% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh.MAD.R
\name{oh.MAD}
\alias{oh.MAD}
\title{Performs MAD rooting of a phylogenetic tree}
\usage{
oh.MAD(unrooted_newick, output_mode)
}
\description{
Performs MAD rooting of a phylogenetic tree
}
\examples{
oh.MAD()
}
\keyword{MAD,}
\keyword{phylogenetics}
\keyword{tree,}
