% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh.save.pdf.R
\name{oh.save.pdf}
\alias{oh.save.pdf}
\title{Function to save objects in high resolution. To be used with illustrator}
\usage{
oh.save.pdf(object = NULL, outname = NULL, outdir = "figures/",
  width = 20, height = 15, family = "Arial",
  fallback_resolution = 1200, antialias = "default")
}
\description{
Pass any type of figure
}
\examples{
oh.save.pdf()
}
\keyword{save,figure,illustrator,vectorized}
