% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ohchibi.R
\name{theme_ohchibi}
\alias{theme_ohchibi}
\title{ggplo2 theme default for ohchibi plots}
\usage{
theme_ohchibi(size_axis_text = 20, size_axis_title = 30,
  legend_proportion_size = 4, size_legend_text = 20,
  size_lines_panel = 0.3, size_panel_border = 1, x_hjust = 0.5,
  y_vjust = 0.5)
}
\description{
Generalized theme for ggplot2. As universal as possible
}
\examples{
theme_ohchibi()
}
\keyword{ggplot2,}
\keyword{theme}
